pyproject.toml
[tool.poetry]
name = "summon_python"
version = "0.1.0"
description = ""
authors = ["Tarcísio Eduardo Moreira Crocomo <tarcisioe@pm.me>"]

[tool.poetry.dependencies]
python = "^3.7"
toml = "^0.10.2"
summon-tasks = { version = "^0.1.0", optional = true }

[tool.poetry.plugins.summon]
python = "summon_python.plugin"

[tool.poetry.dev-dependencies]
types-toml = "^0.10.7"
mypy = "^0.950"
pylint = "^2.13.9"
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"
flake8-bugbear = "^22.4.25"
pylint-quotes = "^0.2.3"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"

[tool.black]
skip_string_normalization = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.summon.plugins.python]
test-modules = ['tests']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
